Add command for ingesting food truck data
Adds a command for loading data for food trucks. Since the data seems to
not change that often, stores the responses Etag to determine whether it
needs to reload all of the data. The data is relatively small, so we're
just truncating the old data and repopulating it.

This command should be run with Laravel's scheduler or via cron,
whichever you prefer.